+++ scripts/travis/setup_database.sh
@@ -4,3 +4,4 @@ set -ev
 
 psql -c 'CREATE DATABASE catmaid;' -U postgres
 psql -c 'CREATE EXTENSION postgis;' -U postgres catmaid
+psql -c 'CREATE EXTENSION pg_trgm;' -U postgres catmaid